requirements.txt
# --- Core scientific stack
numpy>=1.24
pandas>=2.0
scipy>=1.10
tqdm>=4.66

# --- Plotting/HTML viz
matplotlib>=3.7
plotly>=5.20

# --- NLP utilities
nltk>=3.8
langdetect>=1.0.9
tiktoken>=0.7
vaderSentiment>=3.3.2

# --- Embeddings / vector search (RAG + BERTopic)
torch>=2.1           # CPU wheels install automatically via pip
sentence-transformers>=2.6
faiss-cpu>=1.7.4     # use 'faiss-cpu' on Windows/Linux/macOS (CPU)

# --- Topic modeling stack
bertopic>=0.16
umap-learn>=0.5.5
hdbscan>=0.8.33
scikit-learn>=1.3

# --- LLM providers (used by llm_config.py)
openai>=1.30
mistralai>=0.4.0
transformers>=4.41
huggingface-hub>=0.23

# --- Infra / helpers
python-dotenv>=1.0
requests>=2.31
